requirements.txt
moviepy==1.0.3
pytrends==4.9.3
pyttsx3==2.90
Pillow==9.5.0
google-api-python-client==2.114.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
numpy==1.24.4
